Refactor: remove `vocab::data` module, restructure and simplify vocab hierarchy, and update all imports and dependencies accordingly.
