+++ crates/bpetok/src/tokenizer/unified_encoder.rs
@@ -6,8 +6,8 @@ use crate::tokenizer::TokenEncoder;
 use crate::types::TokenType;
 use crate::util::regex::regex_wrapper::RegexWrapper;
 use crate::util::validators;
-use crate::vocab::data::TokenVocab;
-use crate::vocab::data::unified_vocab::UnifiedTokenVocab;
+use crate::vocab::TokenVocab;
+use crate::vocab::unified_vocab::UnifiedTokenVocab;
 use std::sync::Arc;
 
 /// Config options for the [`ScanningEncoder`].
@@ -168,9 +168,9 @@ mod tests {
     use crate::decoder::TokenDecoder;
     use crate::tokenizer::TokenEncoder;
     use crate::tokenizer::unified_encoder::ScanningEncoder;
+    use crate::training::trainer::{BPETokenVocabTrainer, TrainResults};
     use crate::types::{check_is_send, check_is_sync};
-    use crate::vocab::data::unified_vocab::UnifiedTokenVocab;
-    use crate::vocab::training::trainer::{BPETokenVocabTrainer, TrainResults};
+    use crate::vocab::unified_vocab::UnifiedTokenVocab;
     use compact_str::CompactString;
     use std::sync::Arc;
