+++ crates/bpetok/src/training/word.rs
