+++ crates/bpetok/src/training/word_count.rs
@@ -1,7 +1,7 @@
 //! # Text Splitting
 
+use crate::training::word::Word;
 use crate::types::{CountType, StringChunkType, TokenType};
-use crate::vocab::training::word::Word;
 use ahash::AHashMap;
 use std::fmt::Debug;
 
@@ -250,8 +250,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::training::word::Word;
     use crate::types::{CountType, StringChunkType};
-    use crate::vocab::training::word::Word;
     use compact_str::CompactString;
 
     const PATTERN: &str = r"\w+";
